% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bidimensional.R
\name{lm2}
\alias{lm2}
\title{Fitting Bidimensional Regression Models}
\usage{
lm2(formula, data, transformation)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted in the format \code{A + B ~ C + D}, where
\code{A} and \code{B} are dependent and \code{C} and \code{D} are indepdent variables}

\item{data}{a data frame containing variables for the model.}

\item{transformation}{the transformation to be used, either \code{'euclidean'} or \code{'affine'}.}
}
\value{
lm2 returns an object of class "lm2".
}
\description{
lm2 is used to fit bidimensional linear regression models using
Euclidean and Affine transformations following the approach by Tobler (1965).
}
\examples{
nakayaAffine <- lm2(depV1 + depV2 ~ indepV1 + indepV2, NakayaData, 'Affine')
}
